fix(compiler): reparse empty file
